# Update check-elb-latency.rb to AWS SDK v2