**Disclosure: I'm not a Ruby programmer, I cannot run the test suite, please review my changes.**

## Pull Request Checklist

Issue related: https://github.com/sensu-plugins/sensu-plugins-aws/issues/240

#### General

- [X] Update Changelog following the conventions laid out [here](https://github.com/sensu-plugins/community/blob/master/HOW_WE_CHANGELOG.md)

- [ ] Update README with any necessary configuration snippets

- [ ] Binstubs are created if needed

- [ ] RuboCop passes

- [ ] Existing tests pass

#### Purpose

Update to AWS SDK v2.

#### Known Compatibility Issues

This is a breaking change because it removes `aws_access_key` and `aws_secret_access_key` options per #2.

#### Examples

**Before this PR**

Check latency of all ELB in a region having 1 LB
```
$ ./check-elb-latency.rb -a <key> -k <secret> -r eu-west-1
CheckELBLatency OK: <AWS::ELB::LoadBalancer name:my_lb>; (Average within 60 seconds between 2018-10-28 09:12:18 +0000 to 2018-10-28 09:13:18 +0000)
```

Check latency of all ELB in a region having 1 LB, with a warning status
```
$ ./check-elb-latency.rb -a <key> -k <secret> -r eu-west-1 --warning-over 0.1
CheckELBLatency WARNING: <AWS::ELB::LoadBalancer name:my_lb>;  Latency is 0.139 seconds. (expected lower than 0.100); (Average within 60 seconds between 2018-10-28 09:13:35 +0000 to 2018-10-28 09:14:35 +0000)
```

Check latency of all ELB in a region having 2 LB
```
$ ./check-elb-latency.rb -a <key> -k <secret> -r eu-west-1
CheckELBLatency OK: 2 load balancers total; (Average within 60 seconds between 2018-10-28 09:05:11 +0000 to 2018-10-28 09:06:11 +0000)
```

Check latency of all ELB in a region having 2 LB, with a warning status
```
$ ./check-elb-latency.rb -a <key> -k <secret> -r eu-west-1 --warning-over 0.1
CheckELBLatency WARNING: 2 load balancers total; <AWS::ELB::LoadBalancer name:my_lb>'s Latency is 0.138 seconds. (expected lower than 0.100); (Average within 60 seconds between 2018-10-28 09:05:24 +0000 to 2018-10-28 09:06:24 +0000)
```

**After this PR**

Check latency of all ELB in a region having 1 LB
```
$ ./check-elb-latency.rb -r eu-west-1
CheckELBLatency OK: my_lb; (Average within 60 seconds between 2018-10-28 09:14:40 UTC to 2018-10-28 09:15:40 UTC)
```

Check latency of all ELB in a region having 1 LB, with a warning status
```
$ ./check-elb-latency.rb -r eu-west-1 --warning-over 0.1
CheckELBLatency WARNING: my_lb;  Latency is 0.171 seconds. (expected lower than 0.100); (Average within 60 seconds between 2018-10-28 09:15:02 UTC to 2018-10-28 09:16:02 UTC)
```

Check latency of all ELB in a region having 2 LB
```
$ ./check-elb-latency.rb -r eu-west-1
CheckELBLatency OK: 2 load balancers total; (Average within 60 seconds between 2018-10-28 09:08:56 UTC to 2018-10-28 09:09:56 UTC)
```

Check latency of all ELB in a region having 2 LB, with a warning status
```
$ ./check-elb-latency.rb -r eu-west-1 --warning-over 0.1
CheckELBLatency WARNING: 2 load balancers total; my_lb's Latency is 0.204 seconds. (expected lower than 0.100); (Average within 60 seconds between 2018-10-28 09:09:48 UTC to 2018-10-28 09:10:48 UTC)
```