@boutetnico 2018-10-28
@majormoses please see my update based on your feedback and examples in first post. Thank you.