@majormoses 2018-11-02
released: https://rubygems.org/gems/sensu-plugins-aws/versions/15.0.0

I had to manually release because we have clogged up travis by enabling https://dependabot.com to help us with a lot of the maintenance work involved in keeping this project working so I can spend more time adding features and fixing bugs.